requirements.txt
fdk
oci
psycopg2-binary
